elitefinder-painel/ai_service/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.10.0
python-dotenv>=1.0.1
google-generativeai>=0.3.2
openai>=1.12.0
anthropic>=0.18.1
httpx>=0.26.0
pydantic-settings>=2.0.0
pandas>=2.1.0
openpyxl>=3.1.0
